requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
numpy
pandas
scikit-learn
xgboost
imbalanced-learn
mlflow
optuna
PyYAML
pymongo
python-dotenv
jupyter
ipykernel

